Permissions, roles, gates: Back-end

+++ app/Http/Controllers/Api/PostController.php
@@ -52,6 +52,8 @@ class PostController extends Controller
 
     public function store(StorePostRequest $request)
     {
+        $this->authorize('posts.create');
+
         if ($request->hasFile('thumbnail')) {
             $filename = $request->file('thumbnail')->getClientOriginalName();
             info($filename);
@@ -69,6 +71,8 @@ class PostController extends Controller
 
     public function update(Post $post, StorePostRequest $request)
     {
+        $this->authorize('posts.update');
+
         $post->update($request->validated());
 
         return new PostResource($post);
@@ -76,6 +80,8 @@ class PostController extends Controller
 
     public function destroy(Post $post)
     {
+        $this->authorize('posts.delete');
+
         $post->delete();
 
         return response()->noContent();
